Revert "Homework7"
