+++ app/src/main/AndroidManifest.xml
@@ -1,16 +1,15 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="ru.lucass.appname">
-    <uses-permission android:name="android.permission.INTERNET"/>
 
     <application
-        android:name=".MovieApplication"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.AndoidAcademy">
+        <activity android:name=".MovieDetailsActivity"></activity>
         <activity
             android:name=".MainActivity"
             android:label="@string/app_name"
