--- app/src/main/java/ru/lucass/appname/MovieApplication.kt
@@ -1,18 +1,0 @@
-package ru.lucass.appname
-
-import android.app.Application
-
-class MovieApplication : Application()   {
-
-    init {
-        instance = this
-    }
-
-    companion object {
-        private var instance: MovieApplication? = null
-
-        fun getInstance() : MovieApplication {
-            return instance!!
-        }
-    }
-}
