+++ app/src/main/java/ru/lucass/appname/MovieDetailsActivity.kt
@@ -1,0 +1,11 @@
+package ru.lucass.appname
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+
+class MovieDetailsActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_movie_details)
+    }
+}
