--- app/src/main/java/ru/lucass/appname/repository/Repository.kt
@@ -1,4 +1,0 @@
-package ru.lucass.appname.repository
-
-interface Repository {
-}
