--- app/src/main/java/ru/lucass/data/Actor.kt
@@ -1,11 +1,0 @@
-package ru.lucass.data
-
-import android.os.Parcelable
-import kotlinx.parcelize.Parcelize
-
-@Parcelize
- data class Actor(
-    val id: Int,
-    val name: String,
-    val picture: String
-) : Parcelable
