--- app/src/main/java/ru/lucass/data/Genre.kt
@@ -1,7 +1,0 @@
-package ru.lucass.data
-
-import android.os.Parcelable
-import kotlinx.parcelize.Parcelize
-
-@Parcelize
-data class Genre(val id: Int, val name: String) : Parcelable
