--- app/src/main/res/layout/fragment_movies_list.xml
@@ -1,72 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:background="@color/background">
-
-    <ImageView
-        android:id="@+id/imageViewLocation"
-        android:layout_width="16dp"
-        android:layout_height="16dp"
-        android:layout_marginTop="12dp"
-        android:layout_marginLeft="16dp"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:srcCompat="@drawable/location"
-        android:scaleType="centerCrop"
-
-        />
-
-    <TextView
-        android:id="@+id/movieTextView"
-        android:layout_width="375dp"
-        android:layout_height="40dp"
-        android:text="Movies list"
-        android:textSize="14sp"
-        android:textFontWeight="900"
-        android:layout_marginTop="11dp"
-        android:layout_marginLeft="8dp"
-        app:layout_constraintStart_toEndOf="@+id/imageViewLocation"
-        app:layout_constraintTop_toTopOf="parent"
-        android:textColor="@color/white">
-    </TextView>
-
-<!--    <ImageView-->
-<!--        android:id="@+id/imageViewMovie"-->
-<!--        android:layout_width="166dp"-->
-<!--        android:layout_height="248dp"-->
-<!--        android:layout_marginTop="27dp"-->
-<!--        android:layout_marginLeft="14dp"-->
-<!--        app:layout_constraintStart_toStartOf="parent"-->
-<!--        app:layout_constraintTop_toBottomOf="@id/movieTextView"-->
-<!--        app:srcCompat="@drawable/movie"-->
-<!--        android:scaleType="centerCrop"-->
-<!--        />-->
-
-<!--    <TextView-->
-<!--        android:layout_width="24dp"-->
-<!--        android:layout_height="24dp"-->
-<!--        android:background="@color/color13_back"-->
-<!--        android:text="13+"-->
-<!--        android:textAlignment="center"-->
-<!--        android:layout_marginTop="8dp"-->
-<!--        android:layout_marginLeft="6dp"-->
-<!--        app:layout_constraintStart_toStartOf="@+id/imageViewMovie"-->
-<!--        app:layout_constraintTop_toTopOf="@+id/imageViewMovie"-->
-<!--        android:textSize="12sp"-->
-<!--        android:textColor="@color/white">-->
-<!--    </TextView>-->
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/rv_films"
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_marginLeft="8dp"
-        app:layoutManager="androidx.recyclerview.widget.GridLayoutManager"
-        app:spanCount="2"
-        app:layout_constraintTop_toBottomOf="@+id/movieTextView"
-        app:layout_constraintStart_toStartOf="parent"
-        />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
