--- app/src/main/res/layout/view_holder_actor.xml
@@ -1,37 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="100dp"
-    android:layout_height="match_parent">
-    <ImageView
-        android:id="@+id/iv_actor_avatar"
-        android:layout_width="80dp"
-        android:layout_height="80dp"
-        android:layout_marginLeft="8dp"
-        android:gravity="center"
-        android:contentDescription="@string/fragment_actors_avatar_content_description"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        tools:src="@tools:sample/avatars"
-        />
-
-    <TextView
-        android:id="@+id/tv_actor_name"
-        android:layout_width="80dp"
-        android:layout_height="34dp"
-        android:layout_margin="8dp"
-        android:textStyle="normal"
-        android:textFontWeight="900"
-        android:lineHeight="20dp"
-        android:textSize="12sp"
-        android:gravity="left"
-        android:textColor="@color/cast"
-        app:layout_constraintTop_toBottomOf="@+id/iv_actor_avatar"
-        app:layout_constraintStart_toStartOf="parent"
-        tools:text="@tools:sample/full_names"
-        />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
