--- app/src/main/res/layout/view_holder_movie.xml
@@ -1,82 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content" >
-
-    <ImageView
-        android:id="@+id/iv_film_avatar"
-        android:layout_width="166dp"
-        android:layout_height="248dp"
-        android:layout_marginLeft="0dp"
-        android:gravity="center"
-        android:contentDescription="@string/fragment_actors_avatar_content_description"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        tools:src="@tools:sample/avatars"
-        />
-
-<!--    <TextView-->
-<!--        android:layout_width="24dp"-->
-<!--        android:layout_height="24dp"-->
-<!--        android:background="@color/color13_back"-->
-<!--        android:text="13+"-->
-<!--        android:textAlignment="center"-->
-<!--        android:layout_marginTop="8dp"-->
-<!--        android:layout_marginLeft="6dp"-->
-<!--        app:layout_constraintStart_toStartOf="@+id/imageViewMovie"-->
-<!--        app:layout_constraintTop_toTopOf="@+id/imageViewMovie"-->
-<!--        android:textSize="12sp"-->
-<!--        android:textColor="@color/white">-->
-<!--    </TextView>-->
-
-    <TextView
-        android:layout_width="24dp"
-        android:layout_height="24dp"
-        android:background="@color/color13_back"
-        android:text="13+"
-        android:id="@+id/tv_PG_Movie"
-        android:textAlignment="center"
-        android:layout_marginTop="8dp"
-        android:layout_marginLeft="6dp"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        android:textSize="12sp"
-        android:textColor="@color/white">
-    </TextView>
-
-    <TextView
-        android:layout_width="146dp"
-        android:layout_height="12dp"
-        android:background="@color/color13_back"
-        android:text="Action"
-        android:id="@+id/tv_TagMovie"
-        android:textAlignment="textStart"
-        android:layout_marginTop="182dp"
-        android:layout_marginLeft="6dp"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/tv_PG_Movie"
-        android:textSize="8sp"
-        android:textColor="@color/tagColor">
-    </TextView>
-
-    <TextView
-        android:layout_width="154dp"
-        android:layout_height="20dp"
-        android:background="@color/color13_back"
-        android:text="Avengers: End Game"
-        android:id="@+id/tv_FilmName"
-        android:textAlignment="center"
-        android:layout_marginTop="224dp"
-        android:layout_marginLeft="6dp"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/tv_PG_Movie"
-        android:textSize="14sp"
-        android:textFontWeight="900"
-        android:textColor="@color/white">
-    </TextView>
-
-</androidx.constraintlayout.widget.ConstraintLayout>
