+++ app/src/main/res/values/colors.xml
@@ -13,5 +13,4 @@
     <color name="disableStar">#6D6D80</color>
     <color name="devastating">#ECECEC</color>
     <color name="cast">#D8D8D8</color>
-    <color name="color13_back">#191926</color>
 </resources>
