+++ app/src/main/res/values/strings.xml
@@ -15,6 +15,4 @@
     <string name="Evans_cast">Chris Evans</string>
     <string name="Ruffalo_cast">Mark Ruffalo</string>
     <string name="Hemswort_cast">Chris Hemsworth</string>
-    <string name="fragment_actors_avatar_content_description">Actor avatar</string>
-    <string name="fragment_films_chosen_text">You chose: %1$s</string>
 </resources>
